+++ tokens/admin.go
@@ -9,7 +9,7 @@ import (
 
 // NewAdminAuthToken generates and returns a new admin authentication token.
 func NewAdminAuthToken(app core.App, admin *models.Admin) (string, error) {
-	return security.NewToken(
+	return security.NewJWT(
 		jwt.MapClaims{"id": admin.Id, "type": TypeAdmin},
 		(admin.TokenKey + app.Settings().AdminAuthToken.Secret),
 		app.Settings().AdminAuthToken.Duration,
@@ -18,7 +18,7 @@ func NewAdminAuthToken(app core.App, admin *models.Admin) (string, error) {
 
 // NewAdminResetPasswordToken generates and returns a new admin password reset request token.
 func NewAdminResetPasswordToken(app core.App, admin *models.Admin) (string, error) {
-	return security.NewToken(
+	return security.NewJWT(
 		jwt.MapClaims{"id": admin.Id, "type": TypeAdmin, "email": admin.Email},
 		(admin.TokenKey + app.Settings().AdminPasswordResetToken.Secret),
 		app.Settings().AdminPasswordResetToken.Duration,
@@ -27,7 +27,7 @@ func NewAdminResetPasswordToken(app core.App, admin *models.Admin) (string, erro
 
 // NewAdminFileToken generates and returns a new admin private file access token.
 func NewAdminFileToken(app core.App, admin *models.Admin) (string, error) {
-	return security.NewToken(
+	return security.NewJWT(
 		jwt.MapClaims{"id": admin.Id, "type": TypeAdmin},
 		(admin.TokenKey + app.Settings().AdminFileToken.Secret),
 		app.Settings().AdminFileToken.Duration,